Fix wrong sign after 160mA

+++ src/INA3221.h
@@ -290,7 +290,7 @@ class INA3221 {
     void setCurrentSumDisable(ina3221_ch_t channel);
 
     // Gets shunt voltage in uV.
-    int16_t getShuntVoltage(ina3221_ch_t channel);
+    int32_t getShuntVoltage(ina3221_ch_t channel);
 
     // Gets warning alert flag.
     bool getWarnAlertFlag(ina3221_ch_t channel);
